requirements.txt
attrs==21.2.0
boto3==1.18.15
botocore==1.21.15
cycler==0.10.0
iniconfig==1.1.1
jmespath==0.10.0
JPype1==1.3.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
packaging==21.0
pandas==1.3.1
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pyathenajdbc==3.0.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
